Updating the | symbol placed between <li> tags to match core Bugzilla.
The old method was invalid markup, as the | symbol was between <li> elements.  This came about because previously these were just inline <a> tags, where separators after all but the last item made sense.  However, this wasn't updated when they were made into <li> tags.

This commit restyles them to match the way that Bugzilla outputs the separator in its built-in menu lists, so they can be styled in the same way.